feat:Olympics - removed use of helmet and using same code as opta widgets
