+++ packages/ts-components/package.json
@@ -69,7 +69,6 @@
     "react": "16.9.0",
     "react-apollo": "2.5.5",
     "react-elastic-carousel": "^0.11.5",
-    "react-helmet-async": "1.0.2",
     "regenerator-runtime": "0.13.7",
     "sanitize-html": "^2.4.0",
     "styled-components": "4.3.2"
