+++ packages/ts-components/src/components/olympics/medal-table/OlympicsMedalTable.stories.tsx
@@ -2,15 +2,12 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 
 import { OlympicsMedalTable } from './OlympicsMedalTable';
-import { HelmetProvider } from 'react-helmet-async';
 import { ArticleHarness } from '../../../fixtures/article-harness/ArticleHarness';
 import { text, select } from '@storybook/addon-knobs';
 
 storiesOf('Typescript Component/Olympics', module)
   .addDecorator((storyFn: () => React.ReactNode) => (
-    <HelmetProvider context={{}}>
-      <ArticleHarness>{storyFn()}</ArticleHarness>
-    </HelmetProvider>
+    <ArticleHarness>{storyFn()}</ArticleHarness>
   ))
 
   .add('Medal Table', () => {
